+++ src/rt_render/rt_render.c
@@ -6,7 +6,7 @@
 /*   By: jaicastr <jaicastr@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/11 17:36:58 by jaicastr          #+#    #+#             */
-/*   Updated: 2026/03/15 14:08:28 by jaicastr         ###   ########.fr       */
+/*   Updated: 2026/03/15 15:46:59 by jaicastr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,13 +27,14 @@ inline t_taggedresult	rt_alloc_imagebuffer(t_RTContext *ctx)
 	*img = (t_RTImg){.width = (int)ctx->display_width,
 		.height = (int)ctx->display_height, .type = MLX_TYPE_XIMAGE,
 		.format = ZPixmap, .gc = 0, .pix = XCreatePixmap(xvar->display,
-			xvar->root, ctx->display_width, ctx->display_height, xvar->depth)
+			xvar->root, (int)ctx->display_width,
+			(int)ctx->display_height, xvar->depth)
 	};
 	ft_memset(img->data, 0x00, ctx->buffersize);
 	ctx->pix_num = ctx->display_height * ctx->display_width;
 	img->image = XCreateImage(xvar->display, xvar->visual, xvar->depth, ZPixmap,
-			0, (void *)img->data, ctx->display_width,
-			ctx->display_height, 32, 0);
+			0, (void *)img->data, (int)ctx->display_width,
+			(int)ctx->display_height, 32, 0);
 	if (!img->image)
 		return (ft_arena_rewind(&ctx->rt_arena, ctx->rewind_render), KO);
 	img->size_line = img->image->bytes_per_line;
